+++ src/main/java/net/cravencraft/betterparagliders/mixins/paragliders/capabilities/ClientPlayerMovementMixin.java
@@ -1,11 +1,14 @@
 package net.cravencraft.betterparagliders.mixins.paragliders.capabilities;
 
+import net.bettercombat.api.MinecraftClient_BetterCombat;
 import net.bettercombat.logic.PlayerAttackProperties;
 import net.cravencraft.betterparagliders.network.ModNet;
 import net.cravencraft.betterparagliders.network.SyncActionToServerMsg;
+import net.minecraft.client.Minecraft;
 import net.minecraft.world.entity.player.Player;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -15,7 +18,8 @@ import tictim.paraglider.impl.movement.PlayerMovement;
 @Mixin(ClientPlayerMovement.class)
 public abstract class ClientPlayerMovementMixin extends PlayerMovement {
     @Shadow public abstract Player player();
-    private int comboCount;
+    @Unique
+    private int comboCount = 0;
 
     public ClientPlayerMovementMixin(Player player) {
         super(player);
@@ -23,7 +27,14 @@ public abstract class ClientPlayerMovementMixin extends PlayerMovement {
 
     @Inject(method = "update", at = @At(value = "HEAD"), remap=false)
     public void update(CallbackInfo ci) {
-        calculateMeleeStaminaCost();
+
+        // Cancels the Better Combat attack if the player is out of stamina.
+        if (!this.player().isCreative() && !this.player().isSpectator() && this.stamina().isDepleted()) {
+            ((MinecraftClient_BetterCombat) Minecraft.getInstance()).cancelUpswing();
+        }
+        else {
+            calculateMeleeStaminaCost();
+        }
     }
 
     /**
@@ -36,8 +47,8 @@ public abstract class ClientPlayerMovementMixin extends PlayerMovement {
         this.comboCount = (currentCombo == 0) ? currentCombo : this.comboCount;
 
         if (currentCombo > 0 && currentCombo != this.comboCount) {
-            this.comboCount = currentCombo;
             ModNet.NET.sendToServer(new SyncActionToServerMsg(this.comboCount));
+            this.comboCount = currentCombo;
         }
     }
 }