+++ src/main/java/net/cravencraft/betterparagliders/mixins/paragliders/stamina/BotWStaminaMixin.java
@@ -1,12 +1,11 @@
 package net.cravencraft.betterparagliders.mixins.paragliders.stamina;
 
-import net.cravencraft.betterparagliders.attributes.BetterParaglidersAttributes;
 import net.cravencraft.betterparagliders.capabilities.StaminaOverride;
-import net.minecraft.resources.ResourceLocation;
-import net.minecraft.world.entity.player.Player;
+import net.cravencraft.betterparagliders.utils.CalculateStaminaUtils;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
@@ -18,8 +17,6 @@ import tictim.paraglider.api.stamina.Stamina;
 import tictim.paraglider.impl.movement.PlayerMovement;
 import tictim.paraglider.impl.stamina.BotWStamina;
 
-import static tictim.paraglider.api.movement.ParagliderPlayerStates.*;
-
 @Mixin(BotWStamina.class)
 public abstract class BotWStaminaMixin implements Stamina, Copy, Serde, StaminaOverride {
 
@@ -27,7 +24,10 @@ public abstract class BotWStaminaMixin implements Stamina, Copy, Serde, StaminaO
 
     @Shadow public abstract int giveStamina(int amount, boolean simulate);
 
+    @Unique
     private int totalActionStaminaCost;
+    @Unique
+    private String currentPlayerState = "";
 
     @Override
     public int getTotalActionStaminaCost() {
@@ -41,31 +41,22 @@ public abstract class BotWStaminaMixin implements Stamina, Copy, Serde, StaminaO
 
     @Inject(at = @At("HEAD"), remap = false, cancellable = true, method = "update")
     private void injectStaminaValues(@NotNull Movement movement, CallbackInfo ci) {
-        PlayerMovement playerMovement = (PlayerMovement) movement;
-        Player player = playerMovement.player();
         PlayerState state = movement.state();
         int recoveryDelay = movement.recoveryDelay();
         int newRecoveryDelay = recoveryDelay;
-        int staminaDelta;
-        ResourceLocation playerState = movement.state().id();
-
-        if (IDLE.equals(playerState)) {
-            staminaDelta = (int) (movement.state().staminaDelta() + player.getAttributeValue(BetterParaglidersAttributes.IDLE_STAMINA_REGEN.get()));
-        } else if (RUNNING.equals(playerState)) {
-            staminaDelta = (int) (movement.state().staminaDelta() + player.getAttributeValue(BetterParaglidersAttributes.SPRINTING_STAMINA_REDUCTION.get()));
-        } else if (SWIMMING.equals(playerState)) {
-            staminaDelta = (int) (movement.state().staminaDelta() + player.getAttributeValue(BetterParaglidersAttributes.SWIMMING_STAMINA_REDUCTION.get()));
-        } else if (UNDERWATER.equals(playerState)) {
-            staminaDelta = (int) (movement.state().staminaDelta() + player.getAttributeValue(BetterParaglidersAttributes.SUBMERGED_STAMINA_REGEN.get()));
-        } else if (BREATHING_UNDERWATER.equals(playerState)) {
-            staminaDelta = (int) (movement.state().staminaDelta() + player.getAttributeValue(BetterParaglidersAttributes.WATER_BREATHING_STAMINA_REGEN.get()));
-        } else {
-            staminaDelta = movement.state().staminaDelta();
-        }
+        int staminaDelta = CalculateStaminaUtils.getModifiedStateChange((PlayerMovement) movement);
 
         if (state.staminaDelta() < 0 || this.totalActionStaminaCost != 0) {
 
             if (!isDepleted()) {
+                if (CalculateStaminaUtils.getAdditionalMovementStaminaCost(state.id().getPath())) {
+
+                    if (!this.currentPlayerState.equals(state.id().getPath())) {
+                        this.totalActionStaminaCost -= staminaDelta;
+                    }
+                    staminaDelta = 0;
+                }
+
                 //TODO: This probably needs to be redone with the triangular numbers formula
                 staminaDelta = (staminaDelta < 0) ? staminaDelta - this.totalActionStaminaCost : -this.totalActionStaminaCost;
                 if (staminaDelta > 0) {
@@ -81,9 +72,13 @@ public abstract class BotWStaminaMixin implements Stamina, Copy, Serde, StaminaO
             if (recoveryDelay > 0) {
                 newRecoveryDelay--;
             }
-            else if (state.staminaDelta() > 0) {
-                giveStamina(state.staminaDelta(), false);
+            else if (staminaDelta > 0) {
+                giveStamina(staminaDelta, false);
             }
+        }
+
+        if (!this.currentPlayerState.equals(state.id().getPath())) {
+            this.currentPlayerState = state.id().getPath();
         }
 
         // Check for draining stamina
@@ -97,8 +92,8 @@ public abstract class BotWStaminaMixin implements Stamina, Copy, Serde, StaminaO
         }
 
         //noinspection DataFlowIssue
-        newRecoveryDelay = Math.max(0, Math.max(newRecoveryDelay, state.recoveryDelay()));
-        if (recoveryDelay!=newRecoveryDelay) {
+        newRecoveryDelay = Math.max(newRecoveryDelay, state.recoveryDelay());
+        if (recoveryDelay != newRecoveryDelay) {
             movement.setRecoveryDelay(newRecoveryDelay);
         }