+++ src/main/java/net/cravencraft/betterparagliders/mixins/paragliders/stamina/ServerBotWStaminaMixin.java
@@ -5,7 +5,6 @@ import net.cravencraft.betterparagliders.network.ModNet;
 import net.cravencraft.betterparagliders.network.SyncActionToClientMsg;
 import net.cravencraft.betterparagliders.utils.CalculateStaminaUtils;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ProjectileWeaponItem;
 import net.minecraftforge.network.PacketDistributor;
 import org.jetbrains.annotations.NotNull;
@@ -14,7 +13,6 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import tictim.paraglider.api.movement.Movement;
-import tictim.paraglider.api.movement.ParagliderPlayerStates;
 import tictim.paraglider.api.vessel.VesselContainer;
 import tictim.paraglider.impl.movement.ServerPlayerMovement;
 import tictim.paraglider.impl.stamina.BotWStamina;
@@ -34,9 +32,6 @@ public abstract class ServerBotWStaminaMixin extends BotWStamina implements Stam
      * stamina to drain will be calculated. If there is an overlap between a skill being used (e.g., rolling)
      * and an attack (e.g., rolling immediately after an attack to cancel the animation), then that will be
      * accounted for in the total stamina cost so that there isn't any uneven amount of stamina drained.
-     *
-     * @param movement
-     * @param ci
      */
     @Inject(at = @At("HEAD"), remap = false, cancellable = true, method = "update")
     private void updateServerSideMovement(Movement movement, CallbackInfo ci) {
@@ -48,10 +43,6 @@ public abstract class ServerBotWStaminaMixin extends BotWStamina implements Stam
             if(player.getUseItem().getItem() instanceof ProjectileWeaponItem ||
                     CalculateStaminaUtils.DATAPACK_RANGED_STAMINA_OVERRIDES.containsKey(player.getUseItem().getItem().getDescriptionId().replace("item.", ""))) {
                 calculateRangedStaminaCost();
-            }
-
-            if (this.isDepleted()) {
-                checkShieldDisable();
             }
         }
     }
@@ -67,8 +58,6 @@ public abstract class ServerBotWStaminaMixin extends BotWStamina implements Stam
 
     /**
      * Calculates the amount of stamina blocking an attack will cost.
-     *
-     * @param blockedDamage
      */
     public void calculateBlockStaminaCostServerSide(float blockedDamage) {
         syncActionStamina(CalculateStaminaUtils.calculateBlockStaminaCost(this.player, blockedDamage));
@@ -84,40 +73,5 @@ public abstract class ServerBotWStaminaMixin extends BotWStamina implements Stam
     private void syncActionStamina(int actionStaminaCost) {
         this.setTotalActionStaminaCost(actionStaminaCost);
         ModNet.NET.send(PacketDistributor.PLAYER.with(() -> serverPlayerMovement.player()), new SyncActionToClientMsg(actionStaminaCost));
-    }
-
-    /**
-     * Checks if the player is currently holding a shield item. If so, then the modifyShieldCooldown method is called
-     * to determine what to do with the shield.
-     */
-    private void checkShieldDisable() {
-
-        if (this.player.getOffhandItem().getItem().getDescriptionId().contains("shield")) {
-            modifyShieldCooldown(this.player.getOffhandItem().getItem());
-        }
-        else if (this.player.getMainHandItem().getItem().getDescriptionId().contains("shield")) {
-            modifyShieldCooldown(this.player.getMainHandItem().getItem());
-        }
-    }
-
-    /**
-     * Disables the shield cooldown UNLESS the player's current stamina is fully depleted. If the stamina is depleted,
-     * then the cooldown is set to the amount of ticks remaining until the player's stamina is fully replenished. Has
-     * some additional checks to ensure the shield cooldown time stays in sync with the stamina replenish time as well.
-     *
-     * @param shieldItem A main hand or offhand shield being held by the player
-     */
-    private void modifyShieldCooldown( Item shieldItem) {
-        if (this.player.getOffhandItem().getItem().getDescriptionId().contains("shield")) {
-            int recoveryRate = ParagliderPlayerStates.RECOVERY_DELAY;
-            int currentRecoveredAmount = this.serverPlayerMovement.stamina().stamina();
-            float cooldownPercentage = player.getCooldowns().getCooldownPercent(shieldItem, 0.0F);
-            int shieldRecoveryDelay = (int) (this.serverPlayerMovement.stamina().maxStamina() * (1 - cooldownPercentage));
-
-            if (shieldRecoveryDelay > currentRecoveredAmount) {
-                player.getCooldowns().addCooldown(shieldItem, (this.serverPlayerMovement.stamina().maxStamina() - currentRecoveredAmount) / recoveryRate);
-            }
-        }
-
     }
 }